moved components to a new folder, moved state and function from Input to App, started transferring from List to App
