+++ code/src/App.js
@@ -1,10 +1,25 @@
 import React, { useState } from 'react';
 
-import ThoughtsList from "./ThoughtsList";
-import ThoughtInput from "./ThoughtInput";
+import ThoughtsList from "./components/ThoughtsList";
+import ThoughtInput from "./components/ThoughtInput";
+import { THOUGHTS_URL } from "./urls";
 
 export const App = () => {
+  const [thoughts, setThoughts] = useState([]);
   const [input, setInput] = useState("");
+
+  const fetchThoughts = () => {
+    fetch(THOUGHTS_URL)
+      .then(res => res.json())
+      .then(data => setThoughts(data.reverse()));
+    // useEffect(() => {
+    //   fetch(THOUGHTS_URL)
+    //     .then(res => res.json())
+    //     // .then(json => console.log(json))
+    //     .then(json => setThoughts(json));
+
+    // }, [])
+  }
 
   const handleInputChange = (newInput) => {
     setInput(newInput);
@@ -17,7 +32,9 @@ export const App = () => {
         inputText={input}
         onInputChange={handleInputChange}
       />
-      <ThoughtsList />
+      <ThoughtsList
+        thoughtsArray={thoughts}
+      />
     </div>
   )
 }
