+++ code/src/components/ThoughtInput.js
