+++ code/src/components/thought-input.css
