+++ code/src/components/thoughts-list.css
