+++ code/src/urls.js
@@ -1,3 +1,1 @@
-const THOUGHTS_URL = "https://happy-thoughts-technigo.herokuapp.com/thoughts";
-
-export default THOUGHTS_URL;
+export const THOUGHTS_URL = "https://happy-thoughts-technigo.herokuapp.com/thoughts";
